requirements.txt
streamlit>=1.25.0
plotly>=5.15.0
pandas>=1.5.3
numpy>=1.24.0
scipy>=1.10.0
yfinance>=0.2.19
requests>=2.28.2
python-dotenv>=1.0.0
upstox-python-sdk>=2.0.0
openpyxl>=3.1.2

# Optional - useful for additional visualizations / stats
# matplotlib>=3.7.1
# seaborn>=0.12.2
